# Shared interner: fix leak of interned strings from scrape cache