--- a/scrape/scrape.go
+++ b/scrape/scrape.go
@@ -275,7 +275,7 @@ func newScrapePool(cfg *config.ScrapeConfig, app storage.Appendable, jitterSeed 
 		// Update the targets retrieval function for metadata to a new scrape cache.
 		cache := opts.cache
 		if cache == nil {
-			cache = newScrapeCache()
+			cache = newScrapeCache(intern.Global)
 		}
 		opts.target.SetMetadataStore(cache)
 
@@ -396,7 +396,7 @@ func (sp *scrapePool) reload(cfg *config.ScrapeConfig) error {
 			oldLoop.disableEndOfRunStalenessMarkers()
 			cache = oc
 		} else {
-			cache = newScrapeCache()
+			cache = newScrapeCache(intern.Global)
 		}
 		var (
 			t       = sp.activeTargets[fp]
@@ -802,14 +802,16 @@ type scrapeCache struct {
 	// string in addDropped().
 	droppedSeries map[string]*uint64
 
-	// seriesCur and seriesPrev store the labels of series that were seen
-	// in the current and previous scrape.
+	// seriesCur and seriesPrev store the ref of series that were seen
+	// in the current and previous scrape based on the hash.
 	// We hold two maps and swap them out to save allocations.
-	seriesCur  map[uint64]labels.Labels
-	seriesPrev map[uint64]labels.Labels
+	seriesCur  map[uint64]*cacheEntry
+	seriesPrev map[uint64]*cacheEntry
 
 	metaMtx  sync.Mutex
 	metadata map[string]*metaEntry
+
+	interner intern.Interner
 }
 
 // metaEntry holds meta information about a metric.
@@ -825,13 +827,17 @@ func (m *metaEntry) size() int {
 	return len(m.help) + len(m.unit) + len(m.typ)
 }
 
-func newScrapeCache() *scrapeCache {
+func newScrapeCache(interner intern.Interner) *scrapeCache {
+	if interner == nil {
+		interner = intern.Global
+	}
 	return &scrapeCache{
 		series:        map[string]*cacheEntry{},
 		droppedSeries: map[string]*uint64{},
-		seriesCur:     map[uint64]labels.Labels{},
-		seriesPrev:    map[uint64]labels.Labels{},
+		seriesCur:     map[uint64]*cacheEntry{},
+		seriesPrev:    map[uint64]*cacheEntry{},
 		metadata:      map[string]*metaEntry{},
+		interner:      interner,
 	}
 }
 
@@ -858,7 +864,7 @@ func (c *scrapeCache) iterDone(flushCache bool) {
 		// that haven't appeared in the last scrape.
 		for s, e := range c.series {
 			if c.iter != e.lastIter {
-				intern.ReleaseLabels(intern.Global, e.lset)
+				intern.ReleaseLabels(c.interner, e.lset)
 				delete(c.series, s)
 			}
 		}
@@ -883,8 +889,7 @@ func (c *scrapeCache) iterDone(flushCache bool) {
 	c.seriesPrev, c.seriesCur = c.seriesCur, c.seriesPrev
 
 	// We have to delete every single key in the map.
-	for k, lset := range c.seriesCur {
-		intern.ReleaseLabels(intern.Global, lset)
+	for k := range c.seriesCur {
 		delete(c.seriesCur, k)
 	}
 }
@@ -898,12 +903,14 @@ func (c *scrapeCache) get(met string) (*cacheEntry, bool) {
 	return e, true
 }
 
-func (c *scrapeCache) addRef(met string, ref uint64, lset labels.Labels, hash uint64) {
-	if ref == 0 {
-		return
-	}
-	intern.InternLabels(intern.Global, lset)
-	c.series[met] = &cacheEntry{ref: ref, lastIter: c.iter, lset: lset, hash: hash}
+func (c *scrapeCache) addRef(met string, ref uint64, lset labels.Labels, hash uint64) *cacheEntry {
+	// The cache entries are used for staleness tracking so even if ref is
+	// 0 we need to track it.
+	intern.InternLabels(c.interner, lset)
+
+	ce := &cacheEntry{ref: ref, lastIter: c.iter, lset: lset, hash: hash}
+	c.series[met] = ce
+	return ce
 }
 
 func (c *scrapeCache) addDropped(met string) {
@@ -919,15 +926,14 @@ func (c *scrapeCache) getDropped(met string) bool {
 	return ok
 }
 
-func (c *scrapeCache) trackStaleness(hash uint64, lset labels.Labels) {
-	intern.InternLabels(intern.Global, lset)
-	c.seriesCur[hash] = lset
+func (c *scrapeCache) trackStaleness(ce *cacheEntry) {
+	c.seriesCur[ce.hash] = ce
 }
 
 func (c *scrapeCache) forEachStale(f func(labels.Labels) bool) {
-	for h, lset := range c.seriesPrev {
-		if _, ok := c.seriesCur[h]; !ok {
-			if !f(lset) {
+	for hash, ce := range c.seriesPrev {
+		if _, ok := c.seriesCur[hash]; !ok {
+			if !f(ce.lset) {
 				break
 			}
 		}
@@ -1051,7 +1057,7 @@ func newScrapeLoop(ctx context.Context,
 		buffers = pool.New(1e3, 1e6, 3, func(sz int) interface{} { return make([]byte, 0, sz) })
 	}
 	if cache == nil {
-		cache = newScrapeCache()
+		cache = newScrapeCache(intern.Global)
 	}
 	sl := &scrapeLoop{
 		scraper:             sc,
@@ -1434,14 +1440,17 @@ loop:
 		}
 
 		if !ok {
+			ce := sl.cache.addRef(mets, ref, lset, hash)
 			if tp == nil {
 				// Bypass staleness logic if there is an explicit timestamp.
-				sl.cache.trackStaleness(hash, lset)
+				sl.cache.trackStaleness(ce)
 			}
-			sl.cache.addRef(mets, ref, lset, hash)
 			if sampleAdded && sampleLimitErr == nil {
 				seriesAdded++
 			}
+		} else if ce.ref != ref {
+			// Update the ref if it was invalidated.
+			ce.ref = ref
 		}
 
 		// Increment added even if there's an error so we correctly report the
@@ -1506,7 +1515,7 @@ func (sl *scrapeLoop) checkAddError(ce *cacheEntry, met []byte, tp *int64, err e
 	switch errors.Cause(err) {
 	case nil:
 		if tp == nil && ce != nil {
-			sl.cache.trackStaleness(ce.hash, ce.lset)
+			sl.cache.trackStaleness(ce)
 		}
 		return true, nil
 	case storage.ErrNotFound: