Ensures that interning and releasing strings are done in sync with the cache being manipulated to prevent duplicate interned references from being created. 

Adds a metric to track how many interned strings exist to detect leaks. Also adds tests and a benchmark to keep track of this more easily. 

```
name              old time/op    new time/op    delta
StringInterner-8     4.23s ± 0%     2.39s ± 0%  -43.51%

name              old alloc/op   new alloc/op   delta
StringInterner-8     421MB ± 0%     418MB ± 0%   -0.73%

name              old allocs/op  new allocs/op  delta
StringInterner-8     4.41M ± 0%     4.41M ± 0%   -0.00%
```